Move CustomCheckbox into root of Components folder, where it is supposed to be
